feat(SDK): ability to choose supported SDKs
The SDK Bridge script is now purely a proxy through to the appropriate
supported SDK which are now scripts that act like the SDK Bridge did
previously.

The only current available SDK script is the SteamVRSDK script which
supports the SteamVR SDK. It inherits from a BaseSDK abstract class
which defines all of the interfaces required by an SDK script for it
to be compatible with the SDK Bridge.

The SDK is chosen by defining the appropriate SDK name in the Unity
player settings within the `Scripting Define Symbols` text box.

The available SDK defines are:

 * SDK_STEAMVR